% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{AgPalik}
\alias{AgPalik}
\title{Silver in the visible, from Palik}
\format{A dielectric reference class with 4 fields
\describe{
  \item{wavelength}{price, in US dollars}
  \item{epsilon}{weight of the diamond, in carats}
  \item{span}{weight of the diamond, in carats}
  \item{epsilon}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
AgPalik
}
\description{
Dielectric dataset corresponding to the values of Palik.
}
\keyword{datasets}

