% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{AuJC}
\alias{AuJC}
\title{Gold in the visible}
\format{
A dielectric reference class with 4 fields
\describe{
  \item{wavelength}{in nm}
  \item{epsilon}{dielectric function}
  \item{span}{range of wavelengths}
  \item{comment}{optional information}
}
}
\source{
\url{/https://refractiveindex.info/}
}
\usage{
AuJC
}
\description{
Dielectric dataset corresponding to the values of Johnson and Christy
}
\keyword{datasets}
