% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.r
\name{clausius_mossotti}
\alias{clausius_mossotti}
\title{clausius_mossotti}
\usage{
clausius_mossotti(alpha = 1, N = 0.1)
}
\arguments{
\item{alpha}{in nm3}

\item{N}{in nm-3}
}
\value{
relative dielectric function
}
\description{
clausius_mossotti
}
\details{
lorentzian polarizability of a molecule
}
\author{
baptiste Auguie
}
\concept{user_level utility}
