% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{constants}
\alias{constants}
\title{Various physical constants}
\format{
list with names cel h hbar ee eps0 Z0 Nav mu0
}
\usage{
constants
}
\description{
Constants dataset
}
\keyword{datasets}
