% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielectric.r
\docType{class}
\name{dielectric-class}
\alias{dielectric-class}
\alias{dielectric}
\title{spline}
\arguments{
\item{...}{optional arguments passed to smooth.spline}
}
\value{
list
}
\description{
spline interpolation of permittivity
}
\details{
spline interpolation of permittivity
}
\section{Methods}{

\describe{
\item{\code{fun(wavelength, ...)}}{returns a continuous function of the wavelength}

\item{\code{permittivity(new.wavelength, ...)}}{predict a single value}

\item{\code{predict(
  sp = NULL,
  range = span,
  n = length(epsilon),
  new.wavelength = NULL,
  ...
)}}{interpolation with splines}

\item{\code{raw(range = span)}}{return the raw data as real numbers}

\item{\code{spline(...)}}{returns a list of splinefun for the real and imaginary parts}
}}

\author{
baptiste Auguie
}
