% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{dielectric2plot}
\alias{dielectric2plot}
\title{dielectric2plot}
\usage{
dielectric2plot(m)
}
\arguments{
\item{m}{data.frame with wavelength and complex epsilon}
}
\value{
long format data.frame
}
\description{
Conversion to long format data.frame for plotting
}
\details{
Conversion to long format data.frame for plotting
}
\author{
baptiste Auguie
}
