% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversions.R
\name{eV2L}
\alias{eV2L}
\title{eV2L}
\usage{
eV2L(energy)
}
\arguments{
\item{energy}{energy in eV}
}
\description{
Unit conversions
}
\details{
Unit conversions
}
\seealso{
Other conversion: \code{\link{L2eV}}; \code{\link{L2w}};
  \code{\link{t2eV}}
}

