\name{L2eV}
\alias{eV2L}
\alias{L2eV}
\alias{L2w}
\alias{t2eV}
\alias{w2L}
\title{L2eV}
\usage{
  L2eV(wavelength)
}
\arguments{
  \item{wavelength}{}

  \item{energy}{}

  \item{time}{}
}
\value{
  converted unit
}
\description{
  Unit conversions
}
\details{
  Unit conversions
}
\author{
  baptiste Auguié
}

