% Generated by roxygen2 (4.0.0): do not edit by hand
\name{L2w}
\alias{L2w}
\title{L2w}
\usage{
L2w(wavelength)
}
\arguments{
  \item{wavelength}{wavelength in m}
}
\description{
Unit conversions
}
\details{
Unit conversions
}
\seealso{
Other conversion: \code{\link{L2eV}}; \code{\link{eV2L}};
  \code{\link{t2eV}}
}

