% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversions.R
\name{t2eV}
\alias{t2eV}
\title{t2eV}
\usage{
t2eV(time)
}
\arguments{
\item{time}{time in s}
}
\description{
Unit conversions
}
\details{
Unit conversions
}
\seealso{
Other conversion: \code{\link{L2eV}}; \code{\link{L2w}};
  \code{\link{eV2L}}
}

