\name{AgPalik}
\alias{AgPalik}
\docType{data}
\title{ AgPalik }
\description{
  Silver dielectric function from Palik
}
\usage{data(AgPalik)}
\source{
  E. D. Palik. Handbook of Optical Constants of Solids, volume I. Academic Press, 1985
  luxpop.com
}
\examples{
data(AgPalik)
str(AgPalik$raw())
AgPalik$comment
demo(AgPalik)
}
\keyword{datasets}
