\name{dielectric2plot}
\alias{dielectric2plot}
\title{dielectric2plot}
\usage{dielectric2plot(m)}
\description{Conversion to long format data.frame for plotting}
\details{Conversion to long format data.frame for plotting}
\value{long format data.frame}
\author{baptiste Auguie}
\arguments{\item{m}{data.frame with wavelength and complex epsilon}}
