% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversions.R
\name{L2eV}
\alias{L2eV}
\title{L2eV}
\usage{
L2eV(wavelength)
}
\arguments{
\item{wavelength}{wavelength in m}
}
\value{
converted unit
}
\description{
Unit conversions
}
\details{
Unit conversions
}
\seealso{
Other conversion: \code{\link{L2w}}; \code{\link{eV2L}};
  \code{\link{t2eV}}
}

