% Generated by roxygen2 (4.0.0): do not edit by hand

\name{t2eV}
\alias{t2eV}
\title{t2eV}
\usage{
  t2eV(time)
}
\arguments{
  \item{time}{time in s}
}
\description{
  Unit conversions
}
\details{
  Unit conversions
}
\seealso{
  Other conversion: \code{\link{eV2L}}, \code{\link{L2eV}},
  \code{\link{L2w}}
}

